requirements/prod.txt
Django==5.0.4
djangorestframework==3.15.1
python-dotenv==1.0.1
